backport to Go 1.2 by stubbing out sync.Pool usage
