+++ .travis.yml
@@ -8,6 +8,7 @@ install:
 script: make clean && make test && make test
 
 go:
+  - 1.2
   - 1.3
   - 1.4
   - tip
